requirements.txt
pytest
pytest-html
selenium
webdriver-manager
configparser
Pillow